Narrow type of reflection `get*Type()` after calling `has*Type()`
