+++ stubs/ReflectionFunctionAbstract.stub
@@ -14,4 +14,19 @@ abstract class ReflectionFunctionAbstract
     public function getAttributes(?string $name = null, int $flags = 0)
     {
     }
+
+	/**
+	 * @phpstan-assert-if-true !null $this->getReturnType()
+	 */
+	public function hasReturnType(): bool
+	{
+	}
+
+	/**
+	 * @phpstan-assert-if-true !null $this->getTentativeReturnType()
+	 */
+	public function hasTentativeReturnType(): bool
+	{
+	}
+
 }
