+++ stubs/ReflectionProperty.stub
@@ -2,10 +2,19 @@
 
 class ReflectionProperty
 {
+
     /**
      * @return list<ReflectionAttribute<object>>
      */
     public function getAttributes(?string $name = null, int $flags = 0)
     {
     }
+
+	/**
+	 * @phpstan-assert-if-true !null $this->getType()
+	 */
+	public function hasType(): bool
+	{
+	}
+
 }
