--- tests/PHPStan/Analyser/nsrt/reflection-type.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace ReflectionTypeTest;
-
-use function PHPStan\Testing\assertType;
-
-function test(
-	\ReflectionProperty $reflectionProperty,
-	\ReflectionFunctionAbstract $reflectionFunctionAbstract,
-	\ReflectionParameter $reflectionParameter
-){
-	assertType('ReflectionType|null', $reflectionProperty->getType());
-	assertType('ReflectionType|null', $reflectionFunctionAbstract->getReturnType());
-	assertType('ReflectionType|null', $reflectionParameter->getType());
-}
